requirements.txt
playwright
pandas
beautifulsoup4
flask
pymongo
python-dotenv
